refactor : onboardingHeader -> header로 폴더명 수정
